Implement Sudoku grid definition types and structs
A Sudoku grid is composed of 9x9 squares.
A Square is a 2 chars string with first row in ["A";"I"] range and col
in ["1";"9"] range.
A Unit is a set of 9 squares, either a row, a column or a 3x3 separated
box.
Each square therefore belongs to exactly 3 units, one of each kind.
Square peers is a set of all other squares that share the same units.
A square value cannot also appear on its peers.